feat: edge-cloud AI decision engine + 15 task scripts + config management + full sync
- decision/: DecisionEngine with dual-model routing (DeepSeek text 90% + Qwen3-VL vision 10%)
- edge/: EdgePipeline (ScreenAnalyzer + LocalReactor + StateCompiler), StateIngestor, StateStore
- memory/: cross-device memory store + retriever + experience compiler
- stream/: StreamManager + StreamController for A/V relay
- billing/: billing routes + schema
- config-manager/: hierarchical config management (global/device/template)
- android edge/: full CV pipeline + ActionExecutor + 11 DeviceAction variants
- assets/scripts/: 15 AutoX v7 task scripts (抖音/快手/微信/小红书)
- dashboard: SystemControlPanel + ConfigManagement (5 pages) + KeyMapVisualizer + mediaQuery/keyboardShortcuts hooks
- docs: DEPLOYMENT.md + DEVELOPMENT_PLAN.md + EDGE_CLOUD_ARCHITECTURE_SPEC.md
- gitignore: exclude Kotlin/C++ build caches

+++ android-client/app/src/main/java/com/phonefarm/client/floating/EdgeSnapHelper.kt
@@ -53,13 +53,13 @@ class EdgeSnapHelper(
         // Check left edge
         if (x < snapThresholdPx) {
             edge = SnapEdge.LEFT
-            offset = -(bubbleWidth - peekWidthPx)
+            offset = -(bubbleWidth - peekWidthPx).toInt()
             snappedX = offset
         }
         // Check right edge
         else if (x > screenWidth - bubbleWidth - snapThresholdPx) {
             edge = SnapEdge.RIGHT
-            offset = screenWidth - peekWidthPx
+            offset = (screenWidth - peekWidthPx).toInt()
             snappedX = offset
         }
